+++ evernote-sdk-ios/3rdParty/KSHTMLWriter/KSXMLWriter.m
@@ -641,7 +641,7 @@ static NSCharacterSet *sCharactersToEntityEscapeWithoutQuot;
 
 - (void)writeAttributes:(KSXMLWriter *)writer;
 {
-  for (int i = 0; i < [_attributes count]; i+=2)
+  for (NSUInteger i = 0; i < [_attributes count]; i+=2)
   {
     NSString *attribute = [_attributes objectAtIndex:i];
     NSString *value = [_attributes objectAtIndex:i+1];
